Reworked page handlers (#2)
=============== Merge Request Body ===============
This practically redoes the entire project, but it's much nicer 
now.

`Router` is no longer a trait, but a struct which will try to route
a requested route to the correct resolver. This is thread-safe;
 however, the process of building a Router is not, so a 
`RouterBuilder` struct handles that. In addition to the details 
of the Router, this structure contains a `Box` of a `Resolver`-
generating function (that is, a constructor), which is the non-
thread-safe part of building the resolver list.

To bind a page, you need to simply pass 
`ServerBuilder#bind` a structure which has the `Resolver`
 trait implemented. This resolver will be the resolver used 
to handle page bindings until another has been added.

This also implies that you can have multiple resolver
instances at the same time for a server!
================================================

* Refactored handlers until it compiled

Now there is a trait which can be implemented, instead
of the code being written inside lambda statements!

* Ported Hello World example

* Ported majority of simple portal example

* Added optional router params, fixed simple portal

* Added better error handling to simple_portal

* Fixed routing tests

* Fixed build problem in simple_portal

* Fixed file browser example

* Added documentation to places missing it

* file browser *actually* works for all folders

This was a bit of a pain to figure out why it wasn't working.